+++ tests/tipbot/_trial_temp/test.log
@@ -1,0 +1,1 @@
+2020-02-01 11:49:40-0500 [-] Log opened.
