+++ tests/tipbot/test_get_info.py
@@ -1,0 +1,12 @@
+import unittest
+
+import helper
+
+try:
+    from unittest.mock import patch, Mock, MagicMock
+except ImportError:
+    from mock import patch, Mock
+
+class mainTestCase(unittest.TestCase):
+
+    helper.botname = "MoneroTip"
